+++ src/test/java/com/bisnode/opa/StartOpaTaskTest.java
@@ -1,12 +1,17 @@
 package com.bisnode.opa;
 
+import com.bisnode.opa.configuration.OpaPluginConvention;
 import org.gradle.api.Project;
+import org.gradle.internal.impldep.org.junit.rules.TemporaryFolder;
 import org.gradle.testfixtures.ProjectBuilder;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import javax.annotation.Nullable;
+import java.io.IOException;
+import java.net.URL;
+import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -33,7 +38,13 @@ public class StartOpaTaskTest {
     }
 
     @Test
-    public void opaPluginStartTaskSavesProcessInExtProperties() {
+    public void opaPluginStartTaskSavesProcessInExtProperties() throws IOException {
+        OpaPluginConvention convention = project.getConvention().getPlugin(OpaPluginConvention.class);
+        Optional<String> opaBinaryPath = getOpaBinaryPath();
+        assertTrue(opaBinaryPath.isPresent());
+        convention.setLocation(opaBinaryPath.get());
+        convention.setSrcDir(getPathToTmpFolder());
+
         StartOpaTask startOpaTask = (StartOpaTask) project.getTasks().getByName("startOpa");
         startOpaTask.startOpa();
 
@@ -41,4 +52,13 @@ public class StartOpaTaskTest {
         assertTrue(object instanceof Process);
     }
 
+    private Optional<String> getOpaBinaryPath() {
+        return Optional.ofNullable(getClass().getClassLoader().getResource("opa")).map(URL::getPath);
+    }
+
+    private String getPathToTmpFolder() throws IOException {
+        TemporaryFolder tmpdir = new TemporaryFolder();
+        tmpdir.create();
+        return tmpdir.getRoot().getAbsolutePath();
+    }
 }