+++ src/test/java/com/bisnode/opa/StopOpaTaskTest.java
@@ -3,17 +3,18 @@ package com.bisnode.opa;
 import org.gradle.api.Project;
 import org.gradle.testfixtures.ProjectBuilder;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import javax.annotation.Nullable;
-
 import java.io.IOException;
+import java.net.URL;
+import java.util.Optional;
+import java.util.concurrent.TimeUnit;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class StopOpaTaskTest {
 
@@ -32,12 +33,13 @@ public class StopOpaTaskTest {
     }
 
     @Test
-    @Disabled("Unit test should not depend on OPA binary - move to functional test or mock")
-    public void taskStopsOpaProcess() throws IOException {
-        Process process = new ProcessBuilder().directory(project.getRootDir()).command("opa", "run", "-s").start();
+    public void taskStopsOpaProcess() throws IOException, InterruptedException {
+        Optional<String> opaBinaryPath = getOpaBinaryPath();
+        assertTrue(opaBinaryPath.isPresent());
+        Process process = new ProcessBuilder().directory(project.getRootDir()).command(opaBinaryPath.get(), "run", "-s").start();
         project.getExtensions().getExtraProperties().set("opaProcess", process);
 
-        assertTrue(process.isAlive());
+        assertFalse(process.waitFor(3, TimeUnit.SECONDS));
 
         StopOpaTask task = (StopOpaTask) project.getTasks().getByName("stopOpa");
         task.stopOpa();
@@ -48,6 +50,10 @@ public class StopOpaTaskTest {
 
         assertNotNull(opaProcess);
         assertFalse(opaProcess.isAlive());
+    }
+
+    private Optional<String> getOpaBinaryPath() {
+        return Optional.ofNullable(getClass().getClassLoader().getResource("opa")).map(URL::getPath);
     }
 
 }